Refactor labapp and nbapp entry points and validate extensions
